Add files via upload
The worksheet has tabs for all the working day and 4 sampling hours for each day. The sub changes the month and day in the name of the tabs, using the values found in the Admin sheet. Sheets 4 to 7 correspond to Monday, 8 to 11 to Tuesday, etc.. The month and day are in the form MMDD without separation and in the middle of the tab name.